requirements.txt
discord.py==1.2.5
psycopg2_binary==2.8.4
discord==1.0.1
psycopg2==2.8.4
python-dotenv==0.10.3
